test: move event tests to a separate file
